日志收集支持一个容器内的多个文件聚合到同一个logstore

+++ pilot/fluentd_wrapper.go
@@ -7,6 +7,7 @@ import (
 	"os/exec"
 	"syscall"
 	"time"
+	"strings"
 )
 
 var fluentd *exec.Cmd
@@ -19,7 +20,11 @@ func StartFluentd() error {
 	}
 
 	log.Info("start fluentd")
-	fluentd = exec.Command("/usr/bin/fluentd", "-c", "/etc/fluentd/fluentd.conf", "-p", "/etc/fluentd/plugins")
+	cmdArgs := []string{"-c", "/etc/fluentd/fluentd.conf", "-p", "/etc/fluentd/plugins"}
+	if strings.ToUpper(os.Getenv("FLUETND_DEBUG")) == "DEBUG" {
+		cmdArgs = append(cmdArgs, "-v")
+	}
+	fluentd = exec.Command("/usr/bin/fluentd", cmdArgs...)
 	fluentd.Stderr = os.Stderr
 	fluentd.Stdout = os.Stdout
 	err := fluentd.Start()
@@ -46,21 +51,25 @@ func shell(command string) string {
 
 func ReloadFluentd() error {
 	if fluentd == nil {
-		err := fmt.Errorf("fluentd have not started")
-		log.Error(err)
-		return err
+		return fmt.Errorf("fluentd have not started")
 	}
-
-	log.Info("reload fluentd")
+	log.Warn("reload fluentd")
 	ch := make(chan struct{})
 	go func(pid int) {
 		command := fmt.Sprintf("pgrep -P %d", pid)
 		childId := shell(command)
+		if childId == "" {
+			//restart: always
+			os.Exit(1)
+			close(ch)
+			return
+		}
+
 		log.Infof("before reload childId : %s", childId)
 		fluentd.Process.Signal(syscall.SIGHUP)
 		time.Sleep(5 * time.Second)
 		afterChildId := shell(command)
-		log.Infof("after reload childId : %s", childId)
+		log.Infof("after reload childId : %s", afterChildId)
 		if childId == afterChildId {
 			log.Infof("kill childId : %s", childId)
 			shell("kill -9 " + childId)
